### 2023-07-03 #5541 @bernadette-eu
Hi @rowlandseymour ! Thanks for starting the review. I understand that the command to generate the checklist is
"[at]editorialbot generate my checklist", by looking at other issues.

Lampros
