### 2023-07-03 bernadette-eu/Bernadette#4 @rowlandseymour
[JOSS Review] News File
https://github.com/bernadette-eu/Bernadette/blob/52fd51ca4468d33584926755907f56fcd90a89e3/NEWS.md?plain=1#L1

The news file should contain something, you already have `cran-comments.md` which can be easily converted.

https://github.com/openjournals/joss-reviews/issues/5612
