### 2023-07-03 bernadette-eu/Bernadette#5 @rowlandseymour
[JOSS Review] Code of Conduct
https://github.com/bernadette-eu/Bernadette/blob/52fd51ca4468d33584926755907f56fcd90a89e3/CODE_OF_CONDUCT.md?plain=1#L1

Nice to see, but is it linked to anywhere or just accessed from the directory list? Maybe you could link to from the contribution guidelines.

https://github.com/openjournals/joss-reviews/issues/5612
