### 2023-07-03 #5612 @editorialbot
[REVIEW]: Bernadette: Bayesian Inference and Model Selection for Stochastic Epidemics in R
**Submitting author:** @bernadette-eu (<a href="http://orcid.org/0000-0002-1291-2192">Lampros Bouranis</a>)
**Repository:** https://github.com/bernadette-eu/Bernadette
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.5
