### 2023-07-03 bernadette-eu/Bernadette#6 @rowlandseymour
[JOSS Review] Default prior values
https://github.com/bernadette-eu/Bernadette/blob/52fd51ca4468d33584926755907f56fcd90a89e3/R/priors.R#L173

Should the default values be shown in the help comments as well as in the code?

[openjournals/joss-reviews/](https://github.com/openjournals/joss-reviews/issues/5612)
