### 2023-07-03 bernadette-eu/Bernadette#7 @rowlandseymour
[JOSS Review] Paper
https://github.com/bernadette-eu/Bernadette/blob/52fd51ca4468d33584926755907f56fcd90a89e3/paper/paper.md?plain=1#L28

This is a good paper for the JOSS journal I think. My only concern is the first sentence of the summary. It makes is sound like this is just for Covid-19, but as far as I can see, it could be for a much wider range of infectious diseases. Might it be worth being a bit broader here?

https://github.com/openjournals/joss-reviews/issues/5612
