### 2023-07-03 bernadette-eu/Bernadette#8 @rowlandseymour
[JOSS Review] Global Variable
https://github.com/bernadette-eu/Bernadette/blob/52fd51ca4468d33584926755907f56fcd90a89e3/R/globals.R#L1

You might want to add a comment here in case a user isn't familiar with Global Variables or is doing something that conflicts with these names.

https://github.com/openjournals/joss-reviews/issues/5612
