### 2023-07-03 #5612 @rowlandseymour
@samhforbes Thanks for starting the review thread, Sam. I found this to be a substation and well presented package. I spent yesterday playing around with it and trying to break various parts. There are lots of nice error messages and automated tests. From what I can see, it is implementing the Bayesian hierarchal model it is supposed to. The paper sets out the rationale and background, though I think the summary could be improved. There are comments throughout the repository. I've raised a couple of minor issues, and when these are done, I'm happy to recommend it be accepted.
