### 2023-07-03 #5612 @bernadette-eu
Thank you @rowlandseymour for your feedback.

I am currently working to address your comments. @samhforbes I was wondering about good practices when dealing with these open issues/ reviewer comments on Github. Is it more preferable to create a PR (i.e. "joss_reviews") perform the changes and then merge with master? Or work directly on the master branch? Looking at other JOSS issues/reviews, I can see the former happening.

@rowlandseymour I will let you know if I have any follow-up clarifications.

Lampros Bouranis
