### 2023-07-03 #5612 @samhforbes
Hi @bernadette-eu this is up to your personal workflow preferences - we see both examples.
