+++ README.Rmd
@@ -26,6 +26,9 @@ This R package is provided for use under the [GPL-3.0](https://www.gnu.org/licen
 Please use the [issue tracker](https://github.com/bernadette-eu/Bernadette/issues) for bug reports and feature requests.
 All contributions are welcome, in particular those that improve the approach or the robustness of the code base. We also welcome additions and extensions to the underlying model either in the form of options or improvements.
 
+## Code of Conduct
+
+Please note that this project is released with a [Contributor Code of Conduct](https://github.com/bernadette-eu/Bernadette/blob/master/CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
 
 ## Overview
 
