### 2023-07-03 bernadette-eu/Bernadette#5 @bernadette-eu
I have created a branch for the joss review.

For this issue, I have added a "Code of Conduct" section in the [README](https://github.com/bernadette-eu/Bernadette/blob/joss_review/README.md) file, linking to the [CODE_OF_CONDUCT](https://github.com/bernadette-eu/Bernadette/blob/master/CODE_OF_CONDUCT.md).
