+++ NEWS.md
@@ -1,9 +1,23 @@
+# Bernadette 1.1.4
+
+### New features
+
+* The end-user is enabled to change the granularity of the effective contact rate within a time-span of seven days.
+
+# Bernadette 1.1.3
+
+### Bug fixes
+
+* Function R/stan_igbm.fit.R: removed code line 'out$stanfit <- suppressMessages(rstan::sampling(stanfit, data = standata, chains = 0))'.
+* Functions R/posterior_.R: added a stopping rule that tries to locate "theta_tilde" in 'names(object)', which can only be found in the object returned by rstan::optimizing(). A message is shown, instructing the user to perform Markov chain Monte Carlo sampling using rstan::sampling() or rstan::vb().
+
 # Bernadette 1.1.2
 
-# NEWS
+### Bug fixes
 
-## Version 1.1.2
+* Debugged functions R/stan_igbm.fit.R, R/stan_igbm.R and R/summary.stangbm.R.
+* Debugged the R/posterior_.R functions.
 
-## Version 1.1.1
+# Bernadette 1.1.1
 
 * Released to CRAN on 2023-05-18
