### 2023-07-03 bernadette-eu/Bernadette#4 @bernadette-eu
Thank you! I have modified the NEWS.md file, following `cran-comments.md`. See [here](https://github.com/bernadette-eu/Bernadette/blob/joss_review/NEWS.md).
