### 2023-07-03 bernadette-eu/Bernadette#8 @bernadette-eu
I have added a comment at the top of the [globals.R](https://github.com/bernadette-eu/Bernadette/blob/joss_review/R/globals.R) to address this.

Thanks!
