### 2023-07-04 bernadette-eu/Bernadette#6 @bernadette-eu
I have improved the `description` field of the help file for priors.R and stated what the default values are in the help comments of the function `handle_prior`. Note that `handle_prior` is not an exported object, therefore the default values are presented in the `description` field.
