### 2023-07-04 bernadette-eu/Bernadette#7 @bernadette-eu
Thank you for your support.

The package is one of the deliverables of a [Marie Curie Action](https://cordis.europa.eu/project/id/101027218) which aimed at the development of novel statistical methodology for the modeling of infectious diseases like COVID-19. While the Bayesian hierarchical model was indeed implemented to COVID-19, your are correct to say that the package is applicable to a wider range of infectious diseases and provides a template for the end-user.

I have modified the Summary section accordingly.
