+++ DESCRIPTION
@@ -28,7 +28,7 @@ Imports:
     Rcpp (>= 1.0.8.3),
     RcppParallel,
     rstantools,
-    rstan (>= 2.21.3),
+    rstan (>= 2.26.0),
     scales,
     stats,
     utils
@@ -36,15 +36,15 @@ Suggests:
     knitr,
     rmarkdown,
     roxygen2,
-    StanHeaders (>= 2.21.0),
+    StanHeaders (>= 2.26.0),
     testthat (>= 3.0.0)
 LinkingTo: 
     BH (>= 1.78.0-0),
     Rcpp (>= 1.0.8.3),
     RcppEigen (>= 0.3.3.9.1),
     RcppParallel (>= 5.0.1),
-    rstan (>= 2.21.3),
-    StanHeaders (>= 2.21.0)
+    rstan (>= 2.26.0),
+    StanHeaders (>= 2.26.0)
 VignetteBuilder: 
     knitr
 Biarch: true
