### 2023-09-22 bernadette-eu/Bernadette#11 @strengejacke
[JOSS review] - Statement of Need
Related to this submission: https://github.com/openjournals/joss-reviews/issues/5612

I think the readme also requires a short statement of need section, which seems to be missing. You can use the section from the paper as template, I'd say.
