### 2023-09-22 bernadette-eu/Bernadette#12 @strengejacke
[JOSS review] - Documentation
This is more a suggestion than a requirement to recommend accept of your submission - have you considered using automatic pkgdown websites for your package documentation in GitHub?

(https://github.com/openjournals/joss-reviews/issues/5612)
