### 2023-09-22 69c9467
Changes required for rstan >= 2.26.0.
