+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: Bernadette
 Title: Bayesian Inference and Model Selection for Stochastic Epidemics
-Version: 1.1.4
+Version: 1.1.5
 Authors@R: 
     c(person(given   = "Lampros", 
              family  = "Bouranis", 
