### 2023-09-22 bernadette-eu/Bernadette#12 @bernadette-eu
This is a great suggestion, thanks! I have added it on my todo list for the next version!
