### 2023-09-25 #5612 @bernadette-eu
> @bernadette-eu I am going to now begin the post-review stage, so there are a few things for you to do alongside the final checks I need to do. This is a nice package, well done!

Hi @samhforbes ,

All changes were made in the [joss_review](https://github.com/bernadette-eu/Bernadette/tree/joss_review) branch. Let me merge with the master branch first.

Thanks,
Lampros Bouranis
