### 2023-09-25 #5612 @samhforbes
Oh - thanks, sorry I missed that. If you haven't already, please check the capitalisation of references in the paper (I see Cori et al for example is all capitalised in the title, whereas others are not).
