+++ paper/paper.bib
@@ -8,15 +8,15 @@
 }
 
 @Manual{bernadette,
-  title  = {Bernadette: Bayesian Inference and Model Selection for Stochastic Epidemics},
+  title  = {{Bernadette: Bayesian inference and model selection for stochastic epidemics}},
   author = {Lampros Bouranis},
   year   = {2023},
-  note   = {R package version 1.1.3},
+  note   = {R package version 1.1.4},
   url    = {https://CRAN.R-project.org/package=Bernadette},
   }
 
 @Misc{epidemia,
-    title  = {{epidemia: Modeling of Epidemics using Hierarchical Bayesian Models}},
+    title  = {{epidemia: Modeling of epidemics using hierarchical Bayesian models}},
     author = {Scott, J. and Gandy, A. and Mishra, S. and Unwin, J. and Flaxman, S. and Bhatt, S.},
     year   = {2020},
     note   = {R package version 1.0.0},
@@ -52,7 +52,7 @@
 }
 
 @Manual{epiestim,
-  title = {{EpiEstim: Estimate Time Varying Reproduction Numbers from Epidemic Curves}},
+  title = {{EpiEstim: Estimate time varying reproduction numbers from epidemic curves}},
   author = {Cori, A.},
   year = {2021},
   note = {R package version 2.2-4},
@@ -62,7 +62,7 @@
 @article{Cori2013,
   author = {Cori, A. and Ferguson, NM. and Fraser, C. and Cauchemez, S.},
   year   = {2013},
-  title  = {{A New Framework and Software to Estimate Time-Varying Reproduction Numbers During Epidemics}},
+  title  = {{A new framework and software to estimate time-varying reproduction numbers during epidemics}},
   journal= {Am. J. Epidemiol.},
   doi    = {10.1093/aje/kwt133},
 }
@@ -80,7 +80,7 @@ title = {{Practical considerations for measuring the effective reproductive numb
 
 @book{brooks,
   author    = {Brooks, S. and Gelman, A. and Jones, G. and Meng, X.},
-  title     = {Handbook of Markov Chain Monte Carlo},
+  title     = {{Handbook of Markov chain Monte Carlo}},
   publisher = {CRC press},
   year      = 2011
 }
