### 2023-09-25 #5612 @samhforbes
Hi @bernadette-eu please add
- the DOI for Carpenter et al (I think https://doi.org/10.18637/jss.v076.i01 is right)
- Volume and pages for Cori et al 2013
