+++ paper/paper.bib
@@ -48,34 +48,38 @@
   year    = {2017},
   volume  = {76},
   number  = {1},
-  pages   = {1--32}
+  pages   = {1--32},
+ doi      = {10.18637/jss.v076.i01},
 }
 
 @Manual{epiestim,
-  title = {{EpiEstim: Estimate time varying reproduction numbers from epidemic curves}},
+  title  = {{EpiEstim: Estimate time varying reproduction numbers from epidemic curves}},
   author = {Cori, A.},
-  year = {2021},
-  note = {R package version 2.2-4},
-  url = {https://CRAN.R-project.org/package=EpiEstim},
+  year   = {2021},
+  note   = {R package version 2.2-4},
+  url    = {https://CRAN.R-project.org/package=EpiEstim},
 }
 
 @article{Cori2013,
-  author = {Cori, A. and Ferguson, NM. and Fraser, C. and Cauchemez, S.},
-  year   = {2013},
-  title  = {{A new framework and software to estimate time-varying reproduction numbers during epidemics}},
-  journal= {Am. J. Epidemiol.},
-  doi    = {10.1093/aje/kwt133},
+  author  = {Cori, A. and Ferguson, N. and Fraser, C. and Cauchemez, S.},
+  year    = {2013},
+  title   = {{A new framework and software to estimate time-varying reproduction numbers during epidemics}},
+  journal = {American Journal of Epidemiology},
+  volume  = {178},
+  number  = {9},
+  pages   = {1505--1512},
+  doi     = {10.1093/aje/kwt133},
 }
 
 @article{gostic,
   author = {Gostic, K. and McGough, L. and Baskerville, E. and Abbott, S. and Joshi, K. and Tedijanto, C. and Kahn, R. and Niehus, R. and Hay, J. and De Salazar, P. and Hellewell, J. and Meakin, S. and Munday, J. and Bosse, N. and Sherrat, K.e and Thompson, R. and White, L. and Huisman, J. and Scire, J. and Bonhoeffer, S. and Stadler, T. and Wallinga, J. and Funk, S. and Lipsitch, M. and Cobey, S.},
-title = {{Practical considerations for measuring the effective reproductive number, Rt}},
+title     = {{Practical considerations for measuring the effective reproductive number, Rt}},
   journal = {{PLoS Computational Biology}},
-  year = {2020},
-  volume = {16},
-  number = {12},
-  pages = {1-21},
-  doi   = {10.1371/journal.pcbi.1008409}
+  year    = {2020},
+  volume  = {16},
+  number  = {12},
+  pages   = {1--21},
+  doi     = {10.1371/journal.pcbi.1008409}
 }
 
 @book{brooks,
@@ -91,21 +95,21 @@ title = {{Practical considerations for measuring the effective reproductive numb
   year    = {2017},
   journal = {{Statistics and Computing}},
   volume  = {27},
-  pages  = {1413-1432},
-  doi    = {10.1007/s11222-016-9696-4}
+  pages   = {1413-1432},
+  doi     = {10.1007/s11222-016-9696-4}
 }
 
 @Misc{psisloo2,
-  title = {{loo: Efficient leave-one-out cross-validation and WAIC for Bayesian models}},
+  title  = {{loo: Efficient leave-one-out cross-validation and WAIC for Bayesian models}},
   author = {Vehtari, A. and Gabry, J. and Magnusson, M. and Yao, Y. and Bürkner, P. and Paananen, T. and Gelman, A.},
-  year = {2023},
-  note = {R package version 2.6.0},
-  url  = {https://mc-stan.org/loo/},
+  year   = {2023},
+  note   = {R package version 2.6.0},
[…]
