### 2023-09-26 #5612 @samhforbes
Hi @bernadette-eu
I can see under the hood I think the citations aren't formatted in typical md style as JOSS needs it.
See https://joss.readthedocs.io/en/latest/submitting.html section on citation.
