+++ paper/paper.md
@@ -12,15 +12,8 @@ authors:
 affiliations:
   - name: "Department of Statistics, Athens University of Economics and Business, Athens, Greece"
     index: 1
-date: "2023-09-25"
+date: 26 September 2023
 bibliography: paper.bib
-output:
-  md_document:
-    preserve_yaml: TRUE
-    variant: "markdown_strict"
-journal: JOSS
-link-citations: yes
-csl: apa.csl
 ---
 
 # Summary
@@ -29,8 +22,8 @@ The Coronavirus Disease 2019 (COVID-19) outbreak caused by SARS-CoV-2
 has led to developments in Bayesian infectious disease modeling,
 allowing modelers to assess the impact of mitigation strategies on
 transmission and to quantify the burden of the pandemic. The
-`Bernadette` package ([Bouranis, 2023](#ref-bernadette)) for the
-open-source R statistical software ([R Core Team, 2023](#ref-rsoft))
+`Bernadette` package [@bernadette] for the
+open-source R statistical software [@rsoft]
 implements a Bayesian evidence synthesis approach to modeling the
 age-specific transmission dynamics of a disease based on daily mortality
 counts. The functionality of `Bernadette` can be used to reconstruct the
@@ -49,22 +42,19 @@ The modeling of disease transmission dynamics is an active research
 area; a list of R packages dedicated to the analysis of epidemics is
 presented in the R Epidemics Consortium website
 (<https://www.repidemicsconsortium.org/>). Among these packages,
-`EpiEstim` ([Cori et al., 2013](#ref-Cori2013); [Cori,
-2021](#ref-epiestim)) offers a data-driven Bayesian framework for the
+`EpiEstim` [@Cori2013; @epiestim] offers a data-driven Bayesian framework for the
 reconstruction of the time-varying reproduction number of a disease
 which represents the number of secondary infections generated by a case
 infected at day *t*. `EpiEstim` uses areal time-series in the form of
 case counts for a given period and population. The incidence of
 infection is assumed to follow a Poisson process with expectation given
-by a renewal equation. The `epidemia` package ([Scott et al.,
-2020](#ref-epidemia)) links observed areal data to latent infections,
+by a renewal equation. The `epidemia` package [@epidemia] links observed areal data to latent infections,
 which are in turn modeled as a self-exciting process tempered by
 time-varying reproduction numbers. It also supports Bayesian multilevel
 models by pooling information across multiple populations.
 
 `Bernadette` complements these packages by implementing the Bayesian
-hierarchical modeling approach described in Bouranis et al.
-([2022](#ref-bouranis)). It links the observed age-stratified mortality
+hierarchical modeling approach described in @bouranis. It links the observed age-stratified mortality
 counts for the population of a given country over time to latent
 infections via an over-dispersed count model. The change in infections
 over time is governed by a deterministic multi-type compartmental model
@@ -84,9 +74,8 @@ these groups. The effective reproduction number is estimated from the
 daily age-stratified mortality counts, accounting for variations in
 transmissibility that are not obvious from reported infection counts.
 While estimation of the uncertainty over the effective reproduction
-number is itself a challenging problem ([Gostic et al.,
-2020](#ref-gostic)), it is propagated naturally via Markov chain Monte
-Carlo ([Brooks et al., 2011](#ref-brooks)).
[…]
