### 2023-09-26 #5612 @bernadette-eu
@samhforbes When rendering my paper.rmd file to paper.md, the references would be indicated by `#ref-...`. I manually updated the paper.md file, hopefully there will be no issues when creating the metadata file.

Please proceed and let me know if anything comes up.

Thanks.
