### 2023-09-26 #5612 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
