### 2023-09-26 #5612 @bernadette-eu
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05612/joss.05612/10.21105.joss.05612.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.05612/joss.05612/10.21105.joss.05612.pdf) 📄 👈

@samhforbes looking at the final proof, the references in for Carpenter and Cori have not been updated. But the respective references in the intermediate proof looks fine. The rest of the document looks good.
