### 2023-09-26 #5612 @bernadette-eu
> Hi @bernadette-eu in the final link from editorial bot these look right to me, or am I missing something? I see volumes and DOIs

Apparently I was looking at a preview download... Checked again, everything looks good!
