+++ paper/paper.md
@@ -83,7 +83,7 @@ Carlo [@brooks].
 visualization, a main function for specifying and fitting the Bayesian
 hierarchical model and five functions for post-processing and
 visualization of posterior model estimates of important epidemiological
-quantities. The Github page for this package contains a detailed
+quantities. The GitHub page for this package contains a detailed
 [README](https://github.com/bernadette-eu/Bernadette/blob/master/README.md)
 file with a description of the modeling framework and the steps involved
 in the workflow.
