### 2023-09-26 #5612 @bernadette-eu
> @bernadette-eu – could you please merge this tiny PR? [bernadette-eu/Bernadette#14](https://github.com/bernadette-eu/Bernadette/pull/14)

Done!
