### 2023-09-26 #5612 @bernadette-eu
@arfon https://joss.theoj.org/papers/10.21105/joss.05612 does not contain the final pdf yet. Regarding closing this review issue, is this something I should do? Cannot seem to find the related button for this.
